2nd_Year/ca298/2023-ca298-jakubs-shop/Lab_12/backend/requirements.txt
asgiref
attrs==22.2.0
certifi
Django
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.3.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.3
names==0.3.0
numpy==1.24.2
pandas==1.5.3
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
six==1.16.0
sqlparse
tzdata
uritemplate==4.1.1
wincertstore==0.2